Connect bond options and codex entries
