+++ templates/chat/chat-codex-entry.hbs
@@ -1,0 +1,6 @@
+<div class="pfu-codex__entry__chat">
+	<img class="profile-img" src="{{entry.img}}" alt="{{entry.name}}">
+	<span class="name">{{entry.name}}</span>
+	<span class="description">{{entry.description}}</span>
+</div>
+
